lf commit: compress: collapse duplication across executor, pagination, flow, and prompt
