+++ .github/workflows/ci.yml
@@ -19,6 +19,14 @@ jobs:
       - name: Run Rust tests
         run: cargo test --all
 
+  python-test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: astral-sh/setup-uv@v4
+      - name: Run Python tests
+        run: uv run pytest python/tests/
+
   e2e-smoke:
     runs-on: ubuntu-latest
     steps:
